backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyttsx3==2.90
SpeechRecognition==3.10.0
wikipedia==1.4.0
pyjokes==0.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pywhatkit==5.4
wolframalpha==5.0.0
pyaudio==0.2.13
PyAutoGUI==0.9.54
psutil==5.9.5
python-decouple==3.8 